Добавить БД

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,12 +1,14 @@
 package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exceptions.ModelNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.model.MPA;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -14,7 +16,7 @@ public class FilmService {
     private FilmStorage filmStorage;
 
     @Autowired
-    public FilmService(FilmStorage filmStorage) {
+    public FilmService(@Qualifier("filmDbStorage") FilmStorage filmStorage) {
         this.filmStorage = filmStorage;
     }
 
@@ -22,12 +24,16 @@ public class FilmService {
         return filmStorage.getFilms();
     }
 
-    public void addFilm(Film film) {
-        filmStorage.addFilm(film);
+    public Film addFilm(Film film) {
+        return filmStorage.addFilm(film);
     }
 
-    public void changeFilm(Film film) {
-        filmStorage.changeFilm(film);
+    public Film getFilmById(long id) {
+        return filmStorage.getFilmById(id);
+    }
+
+    public Film changeFilm(Film film) {
+        return filmStorage.changeFilm(film);
     }
 
     public void like(long id, long userId) {
@@ -35,13 +41,18 @@ public class FilmService {
     }
 
     public void deleteLike(long id, long userId) {
-        filmStorage.deleteLike(id, userId);
+        Film film = getFilmById(id);
+        if(film.getLikes().contains(userId)) {
+            filmStorage.deleteLike(id, userId);
+        } else {
+            throw new ModelNotFoundException("User not found with id " + userId);
+        }
+
     }
 
     public List<Film> getPopularFilms(int count) {
         List<Film> allFilms = getFilms();
         allFilms.sort((o1, o2) -> Integer.compare(o2.getLikes().size(), o1.getLikes().size()));
-        //allFilms.sort(Comparator.comparingInt(o -> o.getLikes().size()));
         if(allFilms.size() <= count) {
             return allFilms;
         } else {
@@ -49,7 +60,31 @@ public class FilmService {
         }
     }
 
-    public Film getFilmById(long id) {
-        return filmStorage.getFilmById(id);
+    public MPA getMpaById(int id) {
+        List<MPA> mpa = getAllMpa();
+        if(mpa.stream().anyMatch(x -> x.getId() == id)) {
+            return filmStorage.getMpaById(id);
+        } else {
+            throw new ModelNotFoundException("MPA not found with id " + id);
+        }
+
+    }
+
+    public List<MPA> getAllMpa() {
+        return filmStorage.getAllMpa();
+    }
+
+    public Genre getGenreById(int id) {
+        List<Genre> genres = getAllGenres();
+        if(genres.stream().anyMatch(x -> x.getId() == id)) {
+            return filmStorage.getGenreById(id);
+        } else {
+            throw new ModelNotFoundException("Genre not found with id " + id);
+        }
+
+    }
+
+    public List<Genre> getAllGenres() {
+        return filmStorage.getAllGenres();
     }
 }
